# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/group/create": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "创建群",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamCreateGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamCreateGroup"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/group/dissolve": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "解散群",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamDissolveGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/group/invite": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "邀请成员入群",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamInviteAccount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamInviteAccount"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/group/list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "获取该账号所有的群聊信息列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamGetGroupList"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/group/members": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "查看所有群成员",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamGetGroupMembers"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamGetGroupMembers"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/group/name": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "通过群名模糊查找群聊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamGetGroupsByName"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足 7003:非群员",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamGetGroupsByName"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/group/quit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "退出群聊",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamQuitGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/group/transfer": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "转让群",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamTransferGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足",
                        "schema": {
                            "$ref": "#/definitions/common.State"
                        }
                    }
                }
            }
        },
        "/api/group/update": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "group"
                ],
                "summary": "更新群信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 账户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求信息",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ParamUpdateGroup"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "1001:参数有误 1003:系统错误 2007:身份不存在 2008:身份验证失败 2009:权限不足",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.State"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/reply.ParamUpdateGroup"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "common.State": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码，0：成功；否则失败",
                    "type": "integer"
                },
                "data": {
                    "description": "数据，失败时返回空"
                },
                "msg": {
                    "description": "状态的具体描述",
                    "type": "string"
                }
            }
        },
...
        "model.SettingGroup": {
            "type": "object",
            "properties": {
                "group_info": {
                    "description": "群组信息",
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.SettingGroupInfo"
                        }
                    ]
                },
                "is_not_disturb": {
                    "description": "是否免打扰",
                    "type": "boolean"
                },
                "is_pin": {
                    "description": "是否 pin",
                    "type": "boolean"
                },
                "is_show": {
                    "description": "是否显示",
                    "type": "boolean"
                },
                "last_show": {
                    "description": "最后显示时间",
                    "type": "string"
                },
                "nick_name": {
                    "description": "昵称（群组时：账号在群中的昵称，好友时：给好友备注的昵称，空表示未设置）",
                    "type": "string"
                },
                "pin_time": {
                    "description": "pin 时间",
                    "type": "string"
                },
                "relation_id": {
                    "description": "关系 ID",
                    "type": "integer"
                },
                "relation_type": {
                    "description": "关系类型['group','friend']",
                    "type": "string"
                }
            }
        },
        "model.SettingGroupInfo": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "群组头像",
                    "type": "string"
                },
                "description": {
                    "description": "群组的描述",
                    "type": "string"
                },
                "name": {
                    "description": "群组名称",
                    "type": "string"
                },
                "relation_id": {
                    "description": "群组 ID",
                    "type": "integer"
                }
            }
        },
...
        "reply.ParamCreateGroup": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "avatar": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "relation_id": {
                    "type": "integer"
                }
            }
        },
...
        "reply.ParamGetGroupList": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SettingGroup"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "reply.ParamGetGroupMembers": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/reply.ParamGroupMemberInfo"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "reply.ParamGetGroupsByName": {
            "type": "object",
            "properties": {
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SettingGroup"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "reply.ParamGroupMemberInfo": {
            "type": "object",
            "properties": {
                "account_id": {
                    "type": "integer"
                },
                "avatar": {
                    "type": "string"
                },
                "is_leader": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                }
            }
        },
...
        "reply.ParamInviteAccount": {
            "type": "object",
            "properties": {
                "invite_member": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "reply.ParamUpdateGroup": {
            "type": "object",
            "required": [
                "avatar",
                "description",
                "name"
            ],
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "request.ParamCreateGroup": {
            "type": "object",
            "required": [
                "description",
                "name"
            ],
            "properties": {
                "description": {
                    "description": "群描述",
                    "type": "string"
                },
                "name": {
                    "description": "群名称",
                    "type": "string"
                }
            }
        },
...
        "request.ParamDissolveGroup": {
            "type": "object",
            "required": [
                "relation_id"
            ],
            "properties": {
                "relation_id": {
                    "description": "要解散的 relationID",
                    "type": "integer"
                }
            }
        },
...
        "request.ParamGetGroupMembers": {
            "type": "object",
            "required": [
                "relation_id"
            ],
            "properties": {
                "page": {
                    "description": "第几页",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页的大小",
                    "type": "integer"
                },
                "relation_id": {
                    "type": "integer"
                }
            }
        },
        "request.ParamGetGroupsByName": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "page": {
                    "description": "第几页",
                    "type": "integer"
                },
                "page_size": {
                    "description": "每页的大小",
                    "type": "integer"
                }
            }
        },
...
        "request.ParamInviteAccount": {
            "type": "object",
            "required": [
                "account_id",
                "relation_id"
            ],
            "properties": {
                "account_id": {
                    "description": "要邀请的群成员",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "relation_id": {
                    "description": "群聊 ID",
                    "type": "integer"
                }
            }
        },
...
        "request.ParamQuitGroup": {
            "type": "object",
            "required": [
                "relation_id"
            ],
            "properties": {
                "relation_id": {
                    "type": "integer"
                }
            }
        },
...
        "request.ParamTransferGroup": {
            "type": "object",
            "required": [
                "relation_id",
                "to_account_id"
            ],
            "properties": {
                "relation_id": {
                    "type": "integer"
                },
                "to_account_id": {
                    "description": "群要转让给的账户ID",
                    "type": "integer"
                }
            }
        },
...
        "request.ParamUpdateGroup": {
            "type": "object"
        },
...
    }
}